Update patch set 2
Patch Set 2:

Hi Niklas,
Sorry -- I actually already cherrypicked this into master and submitted it and it has been part of 3.4 for a while! I just forgot to go and update this external CL.

The way we do development when we open source we push everything from internal out to aosp, so checking it in here doesn't make a difference, but I can do that such that it's more visible in the meantime.

You can see the CL I submitted in October here:
https://android.googlesource.com/platform/tools/base/+/6f7b42cb93dd56d17a28f7a979b854f55c850b37

It's your CL, plus I added a few things, such as a quickfix.

Patch-set: 2
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>