Changed source position toString to output in a more consistent format
Previously if the start line and the end line were the same we would
have the same format as the case where we don't know the end column.
This change prints '?' in case the end column is not not known so it can
be uniquely determined.

Bug: 130838931
Test: updated SourcePositionTest
Change-Id: I928c00d715d9319e289ea557813191cee8006b3a
(cherry picked from commit 34097a9356f54881467b73f6935a151d8d4d544f)