Silence StripDebugSymbolsTask if no .so files to strip
Previously, the StripDebugSymbolsTask would complain if it couldn't find
an NDK, even if there were no .so files to strip.

This CL changes this behavior such that the task only complains if there
are actually .so files to strip.

Fixes: 130901899
Fixes: 132893670
Test: StripDebugSymbolsTaskTest
Change-Id: I2eb9774b81e7fedccd33af8af1b2c60db0dcf4a1
(cherry picked from commit 17948afc85e416433a598afea51d013bf13c1732)