NDK side-by-side: Allow sync to continue when NDK can't be found
We were throwing an exception (due to b/130363042) but we really need
the model in Android Studio so we can give the user hyperlinks that
fix the build.gradle file.

This CL returns the code to the original state of using IssueReporter
to report errors. In order to get full errors reported, the native model
is fully built when the Android model is built since Android Model is
the one that reports errors.

When the bug is fixed we should be able to delay the native model
generation until the native model is requested.

As part of this, I changed NDK auto-download to only happen when the user
explicitly requested a specific NDK version in build.gradle. See comment
in that change for more details.

Bug: 130363042
Test: Reactivate the original test that was disabled.

Change-Id: I1a63d365b9e1eff4236dd008e0488f6b8498ef57
(cherry picked from commit 3dd9d23e2b10b430d8b045ef118f09493c234921)