Fix ConcurrentModificationException on taskSpans.
addSpan() can be called at the same time as writeTaskSpan(), which
causes the backing List (taskSpans) to be modified while it is being
copied by the writer.

Just making a copy of taskSpans before passing it would not be enough,
because we would still be exposed to change the list while we are making
the copy.

Fix: 136028926
Test: SetSdkDirHyperlinkTest not being flaky anymore + existing
Change-Id: I562721238dc0147935376f2acfaf002c1318954d