Support AndroidTest for library using navigation xmls
Previously, ProcessTestManifest would fail for the AndroidTest variant
of library modules with <nav-graph> elements in the AndroidManifest.xml
because the task didn't have access to the navigation json files.

This CL adds the navigation json files as an input to
ProcessTestManifest.

This was never an issue for AndroidTest variants of app modules because
the AndroidManifest.xml is generated in that case, unless the user puts
an AndroidManifest.xml source file in the androidTest source dir.

Bug: 112513722
Test: ManifestMergingTest
Change-Id: If1b8694c3543fff6e58ddfe1a690da66f1c16e05
(cherry picked from commit 94898a4eda63f347dc60ad2f1259316b16521b00)