Convert JetifyTransform to new Gradle APIs
This CL converts JetifyTransform to use new Gradle
APIs, because the old one has been deprecated.

There is a slight change in behavior here, because
now we will copy the input file to output location
even if there is nothing to be rewritten. This small
inefficiency happens because information if a new output
file should be created is known only once Jetifier examines
all class files and returns the result of processing.
Therefore, we must always create an output file, and
consequently, that file must exist after transform is done.
To satisfy this requirement, Jetifier is launched with
a flag that will copy input to output location even if
there are not classes to be rewritten. Additional overhead
is negligible.

BUG: 136294346
Test: existing
Change-Id: I946b5a63450f5218f40b972d767f9b930e330012
(cherry picked from commit ec54b94312d034705eeb5283056caef275ec4ce0)