Fix a nullptr deref when StartTrace fails
This took far to long to track down but once or twice a day we would
end up in a state where StartCapture would crash the transport demon.
The issue was TraceManager::StartProfiling would return null. We would
then attempt to get the config from the nullptr and cause the demon to
crash.

A few good things came out of investigating this.
1) We fixed the transport error reporting.
2) We now have program counter values instead of virtual pointers to
diagnose when a crash in the demon happens.
3) We now have a process to take crashes reported and find the issue.

Test: Updated
Bug: 139791905
Change-Id: I2dd52b1f29956fd67eb38abf0f73154ad7bd3e85
(cherry picked from commit 63ff35e0e3fa3fbadffc2f31cc1ab7dc954cc43d)