Zipflinger: Fix virtual entry writing bug
In its current form, the virtual entry filing algorithm is flawed.
In essence, it tests for space available > 30 and then write 34
bytes. Which is a problem in case where space available is more
than 30 but less than 34.

This does not result in an overflow but rather in a wraparound
and subsequent int -> ushort convertion error if the total size to
fill is a multiple of (65_535 + [30-33]) bytes.

Test: Units Tests
Fixes: 143215332
Change-Id: Ib1140d674812117385fdbbfa02e4ced38482b850