Cache annotation processor detection results
This CL reduces JAVA_PRE_COMPILE and TOTAL_BUILD_TIME for multi-module
projects for version 3.6.

Change-Id I695c89d75b476ce9288ce4db5922bb4d99ecadd8 reduced build times
significantly (for version 4.0) by removing code which was doing
redundant work in multi-module projects.

We can't cherry-pick that CL into 3.6, but this CL improves the
performance of the code deleted in that CL by using caching to reduce
the amount of redundant work done.

Locally, this CL results in a ~3% speedup of TOTAL_BUILD_TIME for
Base100_Java8_codeImplChange (because of a ~30% speedup of
JAVA_PRE_COMPILE)

Bug: 131363875
Test: existing, tested performance manually
Change-Id: Ibc4fba5e63eb8b5e69e3ada9bd7d0b6b211e3e85