Update patch set 1
Patch Set 1:

The only way to transfer information from the Benchmark test to the Benchmark gradle plugin that I discovered is to read from the json file, which is written to the phone by the test, and then this file is copied to the host by the plugin. But after performing the Benchmark test, the json file was missing in the directory displayed in the console by the plugin.

Patch-set: 1
Reviewer: Gerrit User 1458882 <1458882@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>