Update patch set 1
Patch Set 1:

Hey! Thanks for your contribution!

The linked bug is for a separate issue which is asking to fix the cli output of benchmarks (which should probably do it from benchmark code as opposed to in AGP).

Still, removing -1 here is a good fix as it's not supported by all versions of Android, have you tested it against the case where there are multiple files in the additionalTestOutputDir? The logic in reportPathReceiver looks like it expects to process ls output with 1 file per line.

Patch-set: 1
Reviewer: Gerrit User 1458882 <1458882@85c56323-6fa9-3386-8a01-6480fb634889>