Fix compile dependency leak by ConstrainHandler.
Currently, with

 [ Library 1 ]
      |
      |
`implementation`
      |
      V
 [ Library2 ]
      |
      |
`implementation`
      |
      V
 [ library3 ]

[ Library 3 ] will appear in the COMPILE classpath of [ Library 1 ],
when dependency substitution is involved.
At COMPILE time, [ Library 1 ] will be able to use classes defined by
[ Library 3 ], which is not expected.

This happens when [ Library 3 ] is declared as a module dependency by
[ Library 2 ] with

dependency {
    implementation `com.example:library3:1.0`
}

And be substitued with project dependency with

allprojects {
  configurations.all {
      resolutionStrategy.dependencySubstitution {
          substitute module('com.example:library3') with project(':library3')
      }
  }
}

, which is a common usage. See TestWithProjectSubstitution.java.

This is caused by ConstraintHandler, which adds all dependencies from
runtime classpath to compile classpath as dependency constraints, to
make two kinds of classpath consistent with each other.
But when a module dependency from runtime classpath got substituted by
a project dependency, ConstraintHandler will add it to compile classpath
as a REAL dependency instead of a constraint. Thus, the above happens.

This patch fixes the above leak by making ConstraintHandler add a CONSTRAINT
INSTEAD OF a REAL dependency, in the form of a dependency substitution rule,
in the above situation.

This will ACHIEVE the ORIGINAL design INTENT without leak into compile
classpath.
See ConstraintHandler.kt and TestWithProjectSubstitutionCompileDep.java
for detail.

Change-Id: I976e4d736e185762e8e3529d14fdd67615c2e9ae
Signed-off-by: Trol <milestonejxd@gmail.com>