Enable emulator access
This implements go/emu-grpc-integration, which adds a UtpTestRunner that
injects information on how to interact with the emulator gRPC access
point in a safe manner.

Currently it will inject only the grpc.token and grpc.port. TLS support
is slated for a later release.

Includes a set of unit tests.

Bug: 232531109
Test: Series of unit test, sample studio project succeeds
Change-Id: I8910b29efb7d082bce2324a561570406f06edc94