adblib: tweak JDWP session packet IDs
Currently, AS/IJ debugger start at "0" for JDWP packets IDs.
Given adblib needs unique packet IDs for DDMS packets
(`helo`, `apmn`, etc.), it is not a good idea to start at "100".

With this change, we make the start of the packet ID range explicit,
and we use the value of the legacy ddmlib library (0x4000_0000)
for DDMS packets sent on the device JDWP session.

Bug: 241411829
Bug: 188067646
Test: Existing
Change-Id: I5dd3c427fde2e68bd68b3f91c0c3cb9dfa2768b8