Check for process cancellation when merging
This change sprinkles checks liberally throughout the merge process. I
don't have a lot of familiarity with this code, so made best guesses
about where a lot of time might be spent; but any of these checks can
easily be added/removed/moved as need be.

By checking for cancellation frequently, we ensure the read lock being
held can be relinquished as needed to avoid UI freezes.

Bug: 261210448
Test: Manually verified
Change-Id: I1d57224cb212b772aa9f1f634ea5f7db3d80e495