Ensure dexing transform outputs have unique names
Gradle uses the componentId and filename to identify artifacts, but
there can be instances where all dex transform outputs for a given module
have the same filename. This results in a collision in identifiers and
in the DexMergingTask there can be dex files left out of a merge because
gradle only returns a single artifact (even though there should be multiple).

This works around the issue by appending a hash of the input file's absolute path
which ensures each output has a unique name that is still stable between builds.

Gradle issue: https://github.com/gradle/gradle/issues/17213 (may be fixed in 8.0?)
Bug: 246326007

Change-Id: If7aa2a68f35952025d0e1c68cf020a95234e9b9a