Replace the check for hardware accelerated app.
We used to check the layout params of the root view for the
hardware accelerated flag.

It turns out that certain in devices these flags are not consistently
set. I found this was the case with a samsung device from the firebase
device test lab.

The root.isHardwareAccelerated() was correct in the test device.
Switching to use that instead.

Fixes: 244120504
Test: Updated ViewLayoutInspectorTest
Change-Id: I1a7c08f98175f6345fa6c911e9b853ece7be8f95