Use a different util to retrieve annotation
AnnotationUtil.findAnnotation depends on KotlinReferenceProvider*
(as shown at http://gpaste/4923051119476736)
which is migrated to Kotlin repo in order to be part of Analysis APIs.
It is unstable to use that until the migration and version-up of Kotlin
and IJ platform are done. Luckily, there is another util that seems to
work without any visible changes.

Bug: 262044688
Test: existing tests
Change-Id: I70b23ce2897fbb30b11a1662fe88052796dfcca3