Fix testFixtures dependency recognition by ignoring capability version
Do not depend on the version set in the capability when looking for a
testFixtures capability in the resolved variant to figure out if the
dependency is on the testFixtures sourceset. As for project's
testFixtures, the capability contains the version of the project unlike
the external libraries (where the version is set to null).
Having the version as a constraint caused the dependency on the
testFixtures of a versioned project to not be marked as testFixtures
dependency in the model, and so the IDE didn't recognize the dependency
correctly.

Bug: N/A
Test: updated snapshot tests
Change-Id: I7d674c695456e1d79ccef5dcfd04c88e7cab8af1
(cherry picked from commit cd53321adb661222aebe70f4ea961b42d5fba899)