Signflinger: Fix virtual entry with v1 signing
Problem: Signing with v1 results in a virtual entry at the beginning
of the archive.

Solution:
The signer learns about META-INF/MANIFEST.MF properties (e.g:
Created-by, Built-By) by sending a InspectJarEntryRequest seeing our
"outputJarEntry(MANIFEST_ENTRY_NAME)".

This works if the archive already contains a META-INF/MANIFEST.MF but
fails if there is none. The initial approach was to create a
META-INF/MANIFEST.MF, write it to the archive, and let the signer
discover it. The problem with this approach is that it creates a
virtual entry when signing ends. Delete it the first
META-INF/MANIFEST.MF when we receive the final META-INF/MANIFEST.MF is
received (via getAdditionalJarEntries). See (b/268071371).

The workaround is to declare a META-INF/MANIFEST.MF to the signer
*without* writing it to the archive. We then fullfil the signer request
"manually" since it is not in the archive and processRequest would fail
.

Test: V1SigningTest.java
Bug: 268071371
Change-Id: Ic4e9586a771e70bdb3b22927d826d5cc691daf3e