[SDK Index] Change message when multiple issues
... exist for the same library and allow policy issues to be
reported as non-blocking.

When a library has multiple issues we show only one, decided with a
predetermined priority. With this change, we now make it clear that
there are two or more issues instead of picking one.

Also there was an assumption that all policy issues will block
publishing, but that assumption is no longer true; this change uses the
severity from the library for issues that come from the new export.

If the lint message for multiple issues is disabled, then we fall back
to show only one, based on the Policy, Critical (if blocking), Outdated
priority.

Test: GooglePlaySdkIndexTest, GradleDetectorTest
Bug: 284994378
Bug: 284992909
Change-Id: I93f15b69d2eb5f2baacaa76b24c153718f9e914a