Narrow scope of expressions checked by ViewTypeDetector
When checking for unexpected casts, ViewTypeDetector currently treats
the type of the parent of a method call as the target type of a cast,
no matter what that parent is. In b/273250732 the UAST parent of the
findViewById call is a Pair expression, and so a Pair type is incorrectly assumed to be the target type of a cast on the result of the call. Later on, this type also gets incorrectly parsed and the user sees a malformed error.

We should only look for casts where there can be a cast - implicit casts
in assignment statements, and explicit casts from UBinaryExpressionWithType's.

Bug: 273250732
Test: ViewTypeDetectorTest
Change-Id: I9754bf3d2a418fcd021db757f991a1aae755b144