adblib: Improve `ShellCommand.executeAsSingleOutput`
Simplify `executeAsSingleOuput` to use built-in `map {}.first`
flow operators, and stop relying on the custom `firstCollecting`
operator. `firstCollecting` is also deleted in this change, as it
was not used (or that useful) anyways.

Wrt to test: existing test cover most of the code, except the
"device disconnects" code path, which is currently not
supported by FakeAdb. A follow-up CL will add this
capability, and then enable the new test that is marked
as "ignored" in this change.

Bug: 188067646
Test: Existing
Change-Id: Ic5642ba351de6a0d8d0780cc763d29f8ea904fc7