adblib: Fix threading issue
Move a lock one line above. This was caught by a test
that was flaky, so there is no need to add a new test for
this.

We had this sequence:

1. waitForPendingRead
2. Lock
3. Rewind

This means a "read" operation could start between "1" and "2", meaning
"3" (call to `rewind`) could run into a case where there is a pending
read.

Fixes: 290890871
Test: Existing
Change-Id: Ibfc95b8eaae24a6c2fc7f5f7565a43469ae8c16c