[lint] Even more special-case of AGP version checks
The AGP version check in GradleDetector needs to fire on the
pseudo-dependency plugin marker artifacts as well as on
`com.android.tools.build:gradle`; this is particularly visible in the
dependency checking for Version Catalog files.

The version comparison from those marker plugins, and also for the AGP
artifact itself, need to be done with AgpVersion and not Version in
order to handle `-dev` versions "correctly" (where AGP's
interpretation of -dev is "latest preview", while Gradle's is
"earliest in this series").  Implement `maxAgpOrNull` and
`isAgpNewerThan` operators in order to capture those semantics.

Bug: 242691473
Fixes: 295848771
Test: existing test adjusted and strengthened
Change-Id: I184af2fb3e851ecdb1dcfe18c1a7e09705c29213