Enforce a minimum version of the kotlin multiplatform plugin for android
In the new android multiplatform plugin, we are using APIs that were
added across different version of the kotlin multiplatform plugin and
finalized -partially- in 1.9.0. This change enforces the minimum applied
kotlin multiplatform plugin to be 1.9.0 in order for the android target
to work correctly.

This change also separates the interaction between the android plugin
and the kotlin plugin into a separate class, so that when some kotlin
classes are not found on the classpath, the error is more clear that the
kotlin plugin needs to be updated.

Fixes: 295164584
Test: Manual
Change-Id: Ibf66df49a319303f278c9d7d39d8a895a8f6ca73