Improve manifest xml merge  performance during
When merging manifest xmls, we keep the xml node's children in-memory in
form of an immutable list. Every time a new child node is inserted or
deleted, we initialize this list again by going through underlying xml
node's and create a list of wrapper XmlElement objects. This change make
use of LinkedHashMap to preserve the insertion order of the elements and
make the lookup order of magnitude faster. This will only be noticeable
during merging of very large manifest files.

Bug: 298888659
Test: Existing ManifestMerger2Test, ManifestMerger2SmallTest
Change-Id: I5f692339b91b68cd6d7bf4c144cf2f68bd8936b2