Remove obsolete AppWithProvidedLibTest test case
Remove test case which checks for compile-only
AARs. Because AGP is trying to avoid resolving
the runtime classpath during sync, determining
if an AAR is only in the compile classpath is
not possible any more. Of course, we can approximate
this by only examining `compileOnly` configuration,
but there is incorrect as:
- build could be using `runtimeOnly` to add the library
- resolving only dependencies in `compileOnly` does not
necessarily resolve to the same versions as compileClasspath

Therefore, AGP will now emit this warning only during
build, in the AppPreBuildTask.

Bug: 303084082
Test: updating it
Change-Id: I46d6c664693a56dbf430c022ff9781d7494479a6