Add environment-services module
This module is supposed to replace IJ framework bits that are
currently used from intellij-core.jar with a set of services so
that depending on the run environment (in studio, from terminal)
a separate implementation of a service can be injected.

This module starts with a replacement of diagnostics.Logger. A
noop service implementation is provided simplifying the migration
burden for the cases where logging is not important (e.g. tests).

Bug: 306121377
Test: TBD
Change-Id: I94d521827b5058a8800d6d85b8fc88684956d895