Version name picker uses camel case by default
Issue is naming depends on order. If we add first
dependency with name that has hyphen, system will
set it without changes and all future variable will
use hyphen as well.

As we prefer camel case (with suggestions we can see
the whole name) we switch to camel case if the element
is the first one.

Test: existing
Bug: NA
Change-Id: Ibbd9e08ad13b6b3d5332eb239fa139e685ce5b4f