Move minSdkVersionForDexing to CommonBuiltArtifacts
In previous commit 69cda68, AGP passed the minSdkVersionForDexing
property to the IDE *per APK file*. It means that in the case of
multiple split APKs, there will be multiple minSdkVersionForDexing
properties with the same value.

This change updates it so that AGP will now pass one
minSdkVersionForDexing property *per app*. That means in the case of
multple split APKs, there will be only one minSdkVersionForDexing
property.

This makes it easier for the IDE to extract this info from the
output-metadata.json file located next to the APK(s).

Test: Updated GenericBuiltArtifactsTypeAdapterTest
      + New ApkMetadataTest
Bug: 284201412
Change-Id: Ib02acff59f669677917cc16345234634d25ae6f7