Lint: DataFlowAnalyzer: fix "this" expressions
Fix handling of "this" expressions in scope function lambdas. We
previously used a separate UAST visitor, and the expression type or
resolved function to try to work out whether a "this" expression should
be tracked, but unlabelled "this" expressions always refer to the
inner-most receiver, and labelled "this" expressions point to a specific
receiver. We can instead simply visit "this" expressions in our main
visitor (DataFlowAnalyzer), and resolve (via UAST) which receiver is
being referenced. We still need the separate UAST visitor for calls and
simple reference expressions that have an implicit receiver (no explicit
"this" expression).

Bug: 308569204
Test: Updated
Change-Id: If5984c1c751e83b007aeca7d7e33ace4bffa8167