Don't warn in TranslationDetector if something might be wrong
This was already done in one case (when name != originalName), but
we should always return if the resource repository isn't up to date.

I have yet to root-cause this - the resource repository sometimes
(maybe 5% of the time) doesn't contain the resource being visited
when a new resource file is created. I've been able to occasionally repro by moving a resource file from one module to another, into a previously empty resource directory.

Bug: 298647527
Bug: 167583128
Test: manual
Change-Id: I0e97be96e00da1bad78e3e483c8d800940beed11