Add DSL to enable privacy sandbox
To allow us to disable support by default, as it has a small, but
noticable performance penalty.

There is no support in the test only plugin currently, so the flag is
always disabled there, which should not affect any behavior, and when
support is added it should be transparent to the consuming project.

Bug: 310848512
Test: Included API updates
Change-Id: I3cf35e61f56805c3516eb18889b8f900fac2b938