adblib: Close debugger pipeline when session ends
When a JDWP session with an external debugger (e.g.
Android Studio) ends, we need to ensure the JDWP Proxy
and `DebuggerSessionPipeline` terminate normally so that
the debugger socket is closed and released, so that another
debugger session can be started on the same process.

This change also ensure the JDWP Proxy that handles
external (i.e. Android Studio or IntelliJ) debugger connection
requests remains active even across multiple JDWP
connections, even if they fail unexpectedly. The JDWP
proxy need to remain active as long as the JDWP process
is alive.

This change addresses the "detach -> re-attach" use case
that can happen when a user manually connects a debugger
on a running process, ends the debug session, and manually connects again.

This change also cleans up logging related to closing
underlying socket connections and SynchronizedChannel,
so that "expected" exceptions are never logged at the
"info" severity (only "debug").

Bug: 309146676
Test: Included
Change-Id: Ice46e9e5b0c4e37876e8de164a6e4fbfb2aaf024