Depend on tools.common as compileOnly
AGP has a check (checkSameAgpVersion) that essentially checks that
different classloaders (different plugins) do not use multiple
versions of AGP. The check is not very accurate and actually checks
the versions in the com.android.Version class located in
tools.common. Therefore, depending on a version of tools.common
different from the currently used in AGP breaks the project build.
This changes the dependency on tools.common as compile only and in
runtime uses the one that is loaded by AGP itself.

Bug: N/A
Test: manually tested
Change-Id: Ibd00947ef088734dc0e7a8b982e1343f2a7a63e9