Make tools/base build from IDE
As the code shared between tools/base and tools/adt/idea started to use some functionality from guava and mockito absent in the older library versions. The versions in tools/adt/idea were updated while versions in
tools/base were not. This updates the versions of the libraries to the most recent ones.

Bug: N/A
Test: manually tested
Change-Id: Ic70dbcf98aad55be5e55df940d152c6a2bee4234