Fix setting of AndroidVersion.isBaseExtension
We currently assume that if a codename property is set, it's a preview
version, and the current extension level will be the base extension
level. However, a codename of "REL" indicates that this is a release
version, not a preview version.

Update the test to use JUnit4 and non-deprecated assertion methods, and
add a test for API 34-ext10. Compare AndroidVersion values strictly, not
just using equals(), which is too loose for our purposes here.

Bug: 316412991
Test: Updated
Change-Id: I1c7fe5ebeac7043253887ff14adfaf685ca18bab