Use correct locations for TOML statements
In TOML files, the version number declaration and the
library declaration (group+module) are typically in
different places. The GradleDetector was passing the
location marker object for the version declaration
as both the "statement" and "value" locations.

However, for some warnings, we'd like to point to
the library declaration rather than the version
declaration.

Test: Unit tests updated
Bug: 298428496
Change-Id: I73107f8702c79b6b533a8c02cb12f33ddcd47caa