Fix incorrect R8 task dependencies with Android resources disabled
R8Task currently depends on InternalArtifactType.AAPT_PROGUARD_FILE,
but that artifact may not be produced (e.g., when Android resource
processing is disabled).

Gradle has different behaviors when a task consumes a non-existent
artifact set up via ConfigurableFileCollection (adding @Optional
doesn't make a difference):
  + ConfigurableFileCollection.from(null) => OK
  + ConfigurableFileCollection.from(Callable { null }) => OK
  + ConfigurableFileCollection.from(task.project.provider { null })
      => Fails with the following message:
      "Cannot query the value of this property because it has no
      value available"

To fix this issue, we'll let R8Task (which extends
ProguardConfigurableTask) consume AAPT_PROGUARD_FILE only if it is
produced and use a Callable instead of a Provider.

Bug: 319132114
Test: New test in ShrinkLibraryTest
Change-Id: Iedbe41d8b86a2ae9a78af5403a1dff2aaa285c79