adblib: Fix race condition in SerializedSharedFlow
There was a subtle race condition in the `emit` implementation
that required using a `Mutex` to make the operation atomic.

See bug description for details of the issue.

This change includes a new regression test (with minimal
repro steps), and removes a `delay` call in the
`testExecuteGarbageCollectorOnPreApi28DeviceWorks` test that
was used as a workaround for flakiness.

Lack of flakiness was checked by running the bazel commands
below with no failures:

    ./bazel/bazel test //tools/base/adblib-tools:adblib.tools.tests --test_filter=SerializedSharedFlowTest --runs_per_test=1000
    ./bazel/bazel test //tools/base/adblib-ddmlibcompatibility:adblib.ddmlibcompatibility.tests --test_filter=testExecuteGarbageCollectorOnPreApi28DeviceWorks --runs_per_test=1000

Bug: 305098914
Test: Included
Change-Id: I96574e1998634066c923e399612938858bd19715