321010878: Handle library layouts in MissingInflatedId checks
(This is a partial cherry-pick of a bigger fix for Jellyfish.)

In Iguana, AGP changed the way paths are passed to lint's resource
repository, so lint cannot access the source paths of resource
repository items from resources in other modules.

This led to false positives like 321010878.

In Jellyfish, there are bigger fixes in lint for this: in addition
to updating some lint checks with partial analysis to properly
handle this scenario, resource access is locked down further where
it throws errors from detectors attempting to do this, etc.

For Iguana, this CL is only cherry-picking the simplest and safest
part of this: for all detectors that access the paths of resources
in dependent modules, it will only look up resources in the current
module rather than modules with dependencies. That means we'll
have some false negatives, but that's better than false positives.
(Also, for most of these checks, it's generally true that resources
are usually in the same module as the code referencing the resource).

This CL looks like a big change, but it's really just applying
the same simple fix to all these detectors: scope the resource
repository down to just local instead of with dependencies when
in partial analysis mode -- and then tests for all of these.

Test: Unit tests included
Fixes: 321010878
Change-Id: I4e09e000a3fd4469a898f022873af3d940d62e9e