Set `StandardSocketOptions.SO_KEEPALIVE` on adblib socket
Ddmlib is setting `StandardSocketOptions.SO_KEEPALIVE` to
`true` for the socket channel created by a call to `rawExec`.
This change updates the adblib to do the same, but we do it for all
the socket channels with the assumption that setting `SO_KEEPALIVE`
to `true` for a short-lived connection should not adversely
impact a short-lived connection.

Test: existing
Bug: 323822809
Change-Id: I27ec9315ee873ad0974229c7d483376c9d87240c