Include core library desugaring hint in NewApi errors
When API check finds an API violation, it will now also check to see if
this API is part of the core library desugaring backport list, and if
so, it will include that information in the error message. This helps
bring awareness of core library desugaring and offers an alternative way
to deal with violations.

Test: Unit tests included (and updated)
Bug: None
Change-Id: I939793e6864585920d02d91d706eee20c11e1ae2