Add isActivating to TemplateState
If the template is already activating, some providers may not allow
activating it again. This causes the activation action to be disabled,
but the reason for it being disabled is generally of less concern.
Thus, clients may wish to read this flag to see that we're already
activating the template.

Bug: 327729426
Test: n/a
Change-Id: I5784154c989658cafc0fd6d5b13aa29bac908283