Add is_forced_connection Field to DatabaseOpenedEvent
When a forced open database is opened natively by the app, we now send
a `close` event followed by an new `open` event with the `forced` field reset .

Similarly, when an open db that has a forced connection is closed, we again send a `closed` event
followed by an `open` event with the `forced` field set.

This allows Android Studio to know if a database is forced or not.

Note that this might mean that we no longer need the `is_forced_connection` in `GetSchemaResponse` and `QueryResponse`. If this turns out to be the case, they can be removed in a later cl.

Bug: 173312205
Bug: 230166342
Test: Added
Change-Id: I97c4cd06a02daaf719c7cdaf2e7faf8c4e462b9c