add enableCodeCoverage to HostTestBuilder
Since it changes the build flow, it must be set in the
VariantBuilder interfaces. Because the HostTestDslInfo are
created after the HostTestBuilder has been created (to avoid
creating many DslInfo objects if the variant is turned off),
the basic information necessary is added to the main DslInfo
interface.

The CreationConfig interfaces provide the value to be used by
Tasks and TaskManagers.
The CreationConfig/HostTest get the final codeCoverage
value from the HostTestBuilder, therefore it is removed from
the HostTestComponentDslInfo.

Bug: 317492905
Test: Added unit tests and integration tests.
Change-Id: I416fc9e1a3bb7d6f3cca7a0ec4f23ed50c1e0d9f