Make `checkIssueRegistration` more robust against `Issue.create` changes
We were relying on `Issue.create`'s arity to pick the right overloading,
and parameter indices to pick the right parameter, which can be brittle
if/when we add new overloadings of `Issue.create` and deprecate old ones.

This change uses parameter names instead, whose discrepancy is easier to
spot. Performance impact should be minimal in the grand scheme of things.

Bug: n/a
Test: existing
Change-Id: I3885dabaecffd6aca7f467447c93fef7f9734cdf