adblib: process inventory server part 4
In this last part of support the process inventory server,
we add an `ExternalJdwpProcessPropertiesCollector`
extension point to the JDWP process discover.

In addition to JDWP process discovery through
a "regular" JDWP connection to the device,
via the `JdwpProcessPropertiesCollector` class,
we also allow JDWP process discovery through
the process inventory server, via the new
`ExternalJdwpProcessPropertiesCollector`
extension point.

The process inventory server implementation
of `ExternalJdwpProcessPropertiesCollector`
merely collect local updates to a JDWP process
and sends them to the server, while at the same
time, ask the server to sends remote updates
to a JDWP process. This two-way synchronization
eventually leads to all clients of the server to
be up-to-date with JDWP process properties.

Bug: 333388357
Test: Included
Change-Id: If28297ddd199b2f3b8086a1668a6014bfb62f222